Fix float16/float32 dtype mismatch in activation projections
Activations from quantised models are float16 but axes are float32.
Cast activations to float before dot product operations in monitor,
axes, and steering modules.

+++ drift/axes.py
@@ -200,7 +200,7 @@ class AxisManager:
 
                 if target_layer in activations:
                     # Use the mean activation across tokens
-                    act = activations[target_layer].mean(dim=1).squeeze(0)
+                    act = activations[target_layer].mean(dim=1).squeeze(0).float()
                     if is_assistant:
                         assistant_acts.append(act)
                     else:
